+++ utils/label_utils.py
@@ -6,13 +6,13 @@ from custom_types import CategoricalLabels, ComplexMatrix
 
 def cat_to_arg_intervals(
     labels: CategoricalLabels,
-) -> npt.NDArray[np.float_]:
+) -> npt.NDArray[np.float32]:
     """
     Transforms 1D array of categorical labels to 2D array of angle intervals
     Returns a 2D array representing the upper and lower bounds
     of the argument interval for each label
     """
-    new_labels = np.empty(shape=(labels.shape[0], 2), dtype=np.float_)
+    new_labels = np.empty(shape=(labels.shape[0], 2), dtype=np.float32)
 
     categories = np.unique(labels)
 
@@ -30,12 +30,12 @@ def cat_to_arg_intervals(
     return new_labels
 
 
-def cat_to_arg_centers(labels: CategoricalLabels) -> npt.NDArray[np.float_]:
+def cat_to_arg_centers(labels: CategoricalLabels) -> npt.NDArray[np.float32]:
     """
     Splits the unit circle into intervals of equal size based on the number of categories
     Then gives each category a center of one of the intervals
     """
-    new_labels = np.empty_like(labels).astype(np.float_)
+    new_labels = np.empty_like(labels)
 
     cats = np.unique(labels)
 
@@ -47,12 +47,12 @@ def cat_to_arg_centers(labels: CategoricalLabels) -> npt.NDArray[np.float_]:
     return new_labels
 
 
-def cat_to_arg(labels: CategoricalLabels) -> npt.NDArray[np.float_]:
+def cat_to_arg(labels: CategoricalLabels) -> npt.NDArray[np.float32]:
     """
     Samples the 2pi interval for number of unique categories uniformly
     Returns those angles as angles for vectors
     """
-    new_labels = np.empty_like(labels).astype(np.float_)
+    new_labels = np.empty_like(labels).astype(np.float32)
 
     cats = np.unique(labels)
